+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.FunctionHost/Configuration/ChargeIngestionConfiguration.cs
@@ -32,7 +32,7 @@ namespace GreenEnergyHub.Charges.FunctionHost.Configuration
         internal static void ConfigureServices(IServiceCollection serviceCollection)
         {
             serviceCollection.AddScoped<ChargeCommandConverter>();
-            serviceCollection.AddScoped<IHttpFunctionResponseBuilder, HttpFunctionResponseBuilder>();
+            serviceCollection.AddScoped<IHttpResponseBuilder, HttpResponseBuilder>();
             serviceCollection.AddScoped<ValidatingMessageExtractor<ChargeCommandBundle>>();
             serviceCollection.AddScoped<IChargeCommandConverter, ChargeCommandConverter>();
             serviceCollection.AddScoped<SchemaValidatingMessageDeserializer<ChargeCommandBundle>, ChargeCommandDeserializer>();
